bootstrap: breakup `test.rs` into logically contained submodules
